#89: Split reading Granola's local token store out of auth.rs
auth.rs had grown to 601 lines holding two jobs: deciding which token
grans uses, and knowing how to find and decrypt the one Granola's
desktop app left on disk. The second is now api::local_store, leaving
auth.rs as the resolution order and nothing else.

This also matches how the two are expected to age. Local-store
scavenging is a legacy fallback that macOS has already closed off and
Windows may follow; the resolution chain is where new sources get added.

Pure move, no behavior change: 821 tests pass before and after.
